Ex 4.21 - Allow to delete blogs only to blog owner

+++ part4/5-blog/controllers/blogs.js
@@ -63,6 +63,20 @@ blogRouter.put('/:id',async (request,response,next) => {
 })
 
 blogRouter.delete('/:id',async (request,response,next) => {
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, config.SECRET)
+
+    if(!request.token || !decodedToken){
+        return response.status(401).json({error: 'Invalid token'})
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() !== user._id.toString()){
+        return response.status(403).json({error: 'Unauthorized to perform this action'})
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
